requirements.txt
# Dependencias principales
Flask==2.3.3
Flask-Login==0.6.2
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.4
Flask-Caching==2.1.0
requests==2.32.3
PyJWT==2.8.0
redis==5.0.1

# Dependencias de desarrollo
flake8==7.0.0
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
